+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # v0.14.0-dev
 
 * Enhancements
-  * Can now control how claims are added as well as validated
+  * Can now control how claims are added as well as validated by overriding functions in `Joken.Parameters`
   * Add `skip` parameter to `Joken.Token.decode`
 
 * Breaking
-  * `Joken.Codec` is replaced by `Joken.Claims`. 
+  * `Joken.Codec` is replaced by `Joken.Parameters`. 
+  * `json_module` in config is replaced by `parameters_module`. 
 
 # v0.13.0